Moved current code to org.scalatest.autofix.v3_0_6 package, and renamed it as RenameDeprecatedPackage.
