+++ 3.0.6/input/src/main/scala/test/RenameAssertionsForJUnitSpec.scala
@@ -1,5 +1,5 @@
 /*
-rule = ExternalPackagesRename
+rule = RenameDeprecatedPackage
  */
 package test
