+++ 3.0.6/rules/src/main/resources/META-INF/services/scalafix.v1.Rule
@@ -1,1 +1,1 @@
-fix.RenamePackage
+org.scalatest.autofix.v3_0_6.RenameDeprecatedPackage
